moved zoom and camera movement functions to separate files
