+++ include/mouse.hpp
@@ -1,0 +1,4 @@
+#include <SFML/Graphics.hpp>
+
+void updateMousePosView(sf::Vector2i &prevMousePos,sf::Vector2i &currMousePos, sf::RenderWindow &window, sf::View &view);
+void updateMousePosWindow(sf::Vector2i &prevMousePos, sf::Vector2i &currMousePos,sf::RenderWindow &window, sf::View& view);
