+++ include/zoom.hpp
@@ -1,0 +1,3 @@
+#include <SFML/Graphics.hpp>
+
+void zoomView(sf::RenderWindow& window, sf::View& view, int mouseX, int mouseY, float factor);
